requirements.txt
asgiref
attrs==21.4.0
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django
django-filter==21.1
django-money==2.1.1
django-phonenumber-field==6.1.0
djangorestframework==3.13.1
djangorestframework-gis==0.18
drf-yasg==1.20.0
idna==3.3
importlib-metadata==4.11.1
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jsonschema==4.4.0
Markdown==3.3.6
MarkupSafe==2.1.0
packaging==21.3
phonenumbers==8.12.43
psycopg2
py-moneyed==1.2
pyparsing==3.0.7
pyrsistent==0.18.1
pytz==2021.3
PyYAML==6.0
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse
swagger-spec-validator==2.7.4
typing-extensions
uritemplate==4.1.1
urllib3==1.26.8
zipp==3.7.0
